Create staff member

Creates a staff member for use in an academic group website. On the list of all staff members (by default in staff_list.Rmd), the name and description will be reported. In addition, an image, stored in images/staff with filename id.png will be shown.

Usage

staff_member(
  id,
  name,
  description,
  external_link,
  internal_link = FALSE,
  email = NA
)

Arguments

id

String defining the unique identifier for a staff member.

name

Display name of staff member.

description

Short description, used on staff pages.

external_link

Link to an external website.

internal_link

Should an internal link be used? TRUE/FALSE

email

email address for the staff member. Used to pre-fill the templates.

Value

An object of class staff_member with the following elements:

id

name

description

external_link

internal_link

email

Details

The external link will be the link in the "MORE" button on the rendered website.

Examples

example_staff_member <- staff_member(
     id = "example-id",
     name = "Example Name",
     description = "Example description for example staff member.",
     external_link = "http://www.example.com/",
     internal_link = FALSE
 )